+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/engine_functions/logs_database_server_config/logs_database_config_provider.go
@@ -1,0 +1,6 @@
+package logs_database_server_config
+
+type LogsDatabaseConfigProvider interface {
+	//Returns the config content
+	GetConfigContent() (string, error)
+}