+++ container-engine-lib/docs/changelog.md
@@ -6,9 +6,9 @@
 * Added the `ForLogsDatabaseServer` and the `ForLogsCollectorServer` Docker objects attribute
 * Added `LogsCollectorTypeDockerLabelValue`, `ModuleContainerTypeDockerLabelValue` and `LogsDbDataVolumeTypeDockerLabelValue` Docker labels
 * Added `EngineGUIDDockerLabelKey` Docker label key
+* Added `LogsDatabaseConfigProvider` interface for getting logs database config content
 
 # 0.34.0
-
 #### Breaking Changes
 * Move all `docker.DockerKurtosisBackend` functionality to `docker_kurtosis_backend` module
   * Users will need to replace their `github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker` import for